% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eq_china_cleandata}
\alias{eq_china_cleandata}
\title{Clean Data for China Only}
\format{A data frame with 9 rows and 46 variables.}
\source{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
eq_china_cleandata
}
\description{
A clean dataset containing the magnitude and other attributes of earthquakes
in China from the test data set.
}
\keyword{datasets}
