% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadandClean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean and format earthqake data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{a data frame holding the earthquake data}
}
\value{
A data frame with the date and location fields formated correctly
}
\description{
\code{eq_clean_data} creates a date class field and converts location
 coordinates into numeric vectors.
}
\details{
The YEAR field ranges from -2000 or 2000 BC to more contemporary times
 which the usual date functions don't manage well with. This function
 splits the years into AD and BC and creates a DATE field from the
 YEAR, MONTH and DAY fields (most of the MONTH and DAY fields for
 the BC years are NA)
}
\examples{
eq_clean_data(eq_test_data)

}
