% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eq_clean_testdata}
\alias{eq_clean_testdata}
\title{Clean Data from Given Data Set}
\format{A data frame with 100 rows and 46 variables.}
\source{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
eq_clean_testdata
}
\description{
A clean dataset containing the magnitude and other attributes of 100 earthquakes
fromm around the world since recorded time.
}
\keyword{datasets}
