% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create an HTML Label}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{A data frame holding all the information of the earthquake data}
}
\value{
an HTML label for the point on a map.
}
\description{
\code{eq_create_label} Creates a label using HTML in order to label points
on a map.
}
\examples{
eq_create_label(eq_china_cleandata)

}
