% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadandClean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Format the LOCATION_NAME field of the data}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{a data frame holding the earthquake data.}
}
\value{
the earthquake data frame with the LOCATION_NAME field reformatted.
}
\description{
\code{eq_location_clean} reformats the LOCATION_NAME field to exclude
the country name.
}
\details{
Where possible the description up to the colon is stripped out and the
remaining title is set to title case and stored.
}
\examples{
eq_location_clean(eq_test_data)

}
