% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{eq_map}
\alias{eq_map}
\title{Plot Epicentre of Earthquakes on a Map}
\usage{
eq_map(df, annot_col)
}
\arguments{
\item{df}{The data frame holding the longitude and latitudes for the
epicentres of the earthquakes.}

\item{annot_col}{An atomic character holding the field name of the column
to display on the map.}
}
\value{
Return a leaflet map
}
\description{
\code{eq_map} Will use a map to plot the location of the epicentre of an
earthquake. The plot will be done using a circle, with the radius of the
circle indicative of the magnitude of the earthquake.
}
