% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eq_test_data}
\alias{eq_test_data}
\title{Information on 100 Earthquakes}
\format{A data frame with 100 rows and 47 variables. Only 5 will be
described}
\source{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
eq_test_data
}
\description{
A dataset containing the magnitude and other attributes of 100 earthquakes
form around the world since recorded time.
}
\keyword{datasets}
