% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLabellingObject.R
\docType{data}
\name{GeomTimelinelabel}
\alias{GeomTimelinelabel}
\title{Geom class to support plotting geom_timeline}
\description{
Geom class to support plotting geom_timeline
}
\keyword{datasets}
