% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraphs.R
\name{quaketimelinelabel_plot}
\alias{quaketimelinelabel_plot}
\title{Function to plot the time line of earthquakes}
\usage{
quaketimelinelabel_plot(eqdata, x_min, x_max, n_max)
}
\arguments{
\item{eqdata}{a data frame of properly formatted earthquake data.}

\item{x_min}{an atomic character designating the lower end of a date
range.}

\item{x_max}{an atomic character designating the upper end of a date
range.}

\item{n_max}{and atomic numeric designating the number of top earthquakes
by magnitude to label.}
}
\value{
a ggplot2 graphic
}
\description{
\code{quaketimeline_plot} plots the magnitude and number of deaths
of earthquakes on a timeline.
}
\examples{
quaketimelinelabel_plot(eq_china_cleandata, "1800-01-01","2015-01-01", 3)

}
