% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLabellingObject.R
\docType{data}
\name{StatTimelinelabel}
\alias{StatTimelinelabel}
\title{Stat class to support the GeomTimelinelabel class}
\description{
Stat class to support the GeomTimelinelabel class
}
\keyword{datasets}
